modules/accelerators/contextualization/cdf_file_asset_source/functions/fn_dm_extract_assets_by_pattern/requirements.txt
cognite-sdk>=7.83.0,<8.0.0
pyyaml>=6.0
requests
python-dotenv>=1.0.0
pydantic>=2.0.0
cognite-extractor-utils
